Make unit equipment an integer until final output, for easier testing.

+++ src/hoi4_world/countries/hoi4_country.h
@@ -29,7 +29,7 @@ struct NameList
 struct Unit
 {
    std::string unit_template;
-   float equipment;
+   int equipment;
    int location;
 
    std::partial_ordering operator<=>(const Unit&) const = default;
